device-agent/requirements.txt
paho-mqtt
gpiozero 